requirements.txt
django>=4.0,<5.0
django-ninja>=1.0
langgraph>=0.2
groq>=0.4
celery[redis]>=5.3
django-celery-beat>=2.5
python-dotenv>=1.0
requests>=2.31
qrcode[pil]>=7.4
